pyproject.toml
[tool.poetry]
name = "testcontainers-fastapi"
version = "0.1.0"
description = ""
authors = ["Ozenc Kandemir <ozenckandemir@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
testcontainers = "^4.8.2"
httpx = "^0.27.2"
mysql-connector-python = "^9.1.0"
black = "^24.10.0"
ruff = "^0.7.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W"]

[tool.mypy]
files = "."
ignore_missing_imports = true
strict = true
